Added a function for deleting metadata in bulk
Similarly to the bulk addition, we are limited in our ability to use bulk deletion queries by a desire to remain consistent with the API of the original method. Since a count of 0 will return false in that case we are forced to do one-at-a-time deletions in the style of the individual method. There is still some query savings however in the loading of IDs.